@@ -17,9 +17,15 @@
   },
   "repository": "StackStorm/st2flow",
   "engines": {
-    "node": "^8.4.0",
+    "node": "^10.4.0",
     "npm": "^5.3.0"
   },
+  "workspaces": [
+    "modules/*",
+    "../st2web/tasks",
+    "../st2web/modules/st2-test-utils",
+    "."
+  ],
   "browserify": {
     "transform": [
       "babelify",
@@ -62,8 +68,7 @@
   },
   "devDependencies": {
     "@stackstorm/browserify-postcss": "0.3.4-patch.5",
-    "@stackstorm/module-test-utils": "1.0.0-pre.5",
-    "@stackstorm/postcss-less-engine": "0.6.2",
+    "@stackstorm/module-test-utils": "^2.0.0",
     "@stackstorm/st2-build": "1.0.0-pre.5",
     "babel-core": "^6.26.0",
     "babel-eslint": "^8.1.2",
@@ -82,7 +87,7 @@
     "flow-bin": "^0.72.0",
     "gulp": "github:gulpjs/gulp#d8f5c90a0622d19ef1943a2a3d02dc50e3c853e7",
     "ignore-styles": "^5.0.1",
-    "lerna": "^3.0.0-rc.0",
+    "lerna": "^3.4.3",
     "less": "^2.7.3",
     "postcss": "^6.0.14",
     "postcss-import": "11.1.0",
