@@ -1,69 +1,79 @@
 // @flow
 
-import type { TokenRawValue, TokenMapping, TokenCollection, AnyToken, Refinement } from './types';
+import type { JPath, TokenRawValue, TokenKeyValue, TokenMapping, TokenCollection, AnyToken, Refinement } from './types';
 import crawler from './crawler';
 import factory from './token-factory';
+import stringifier from './stringifier';
 
 const DEFAULT_INDENT = '  ';
-const DEFAULT_TAIL = '\n';
 const STR_COLON = ':';
+const STR_DASH = '-';
+const STR_COMMA = ',';
+const STR_OPEN_CURLY = '{';
+const STR_CLOSE_CURLY = '}';
+const STR_OPEN_SQUARE = '[';
+const STR_CLOSE_SQUARE = ']';
+const STR_FIRST_TOKEN = 'mappings.0';
 const REG_INDENT = /\n( +)\S/;
+const REG_LEADING_SPACE = /^\s*/;
 const REG_ALL_WHITESPACE = /^\s+$/;
+const REG_COMMENT = /^\s*#(?:\s*)?/;
+const REG_JSON_START = /^\s*(?:[-:] )?[{[]/;
+const REG_DASH = /^\s*-\s*/;
+const REG_COMMA = /^\s*(?::\s*[{[]\s*)?[}\]]?\s*,\s*/;
 
 /**
  * Class for refining tokens whenever mutations are made to the AST.
  */
 class Refinery {
-  yaml: string;
-  tail: string;
-  head: string;
+  newYaml: string = '';
   indent: string;
+  tree: TokenMapping;
+  lastToken: TokenRawValue;
 
-  constructor(yaml: string, head: string, tail: string) {
-    // The following fields are required for this module to work properly
-    this.yaml = yaml;
-    this.head = head;
-    this.tail = tail;
+  // some stateful properties used while refining the tree
+  jsonDepth: number = -1;
 
-    const match = yaml.match(REG_INDENT);
+  constructor(tree: TokenMapping, oldYaml: string = '') {
+    if(tree.kind !== 2) {
+      throw new Error('Tree argument must be a mapping token');
+    }
+
+    const match = oldYaml.match(REG_INDENT);
     this.indent = match ? match[1] : DEFAULT_INDENT;
+    this.tree = tree;
   }
 
   // This is the only method anybody should care about
-  refineTree(tree: TokenMapping): Refinement {
-    const newTree: TokenMapping = this.prefixToken(tree, 0, tree.jpath);
-    this.reIndexToken(newTree, this.head.length);
+  refineTree(): Refinement {
+    const newTree: TokenMapping = this.prefixToken(this.tree, 0, this.tree.jpath || []);
+    const firstToken: TokenRawValue = crawler.findFirstValueToken(newTree);
+    const startPos: number = firstToken.prefix.reduce((pos, prefix) => pos + prefix.rawValue.length, 0);
+    this.reIndexToken(newTree, startPos);
 
     return {
       tree: newTree,
-      head: this.head,
-      tail: this.tail,
+      yaml: this.newYaml,
     };
   }
 
   /**
    * Given a token, refines the token
-   *
-   * NOTE: this causes side effects on the token!! OMG!!!
    */
-  prefixToken(startToken: AnyToken, depth: number, jpath: Array<string | number>): AnyToken {
+  prefixToken(startToken: AnyToken, depth: number, jpath: JPath): AnyToken {
     startToken.jpath = jpath;
 
     switch(startToken.kind) {
       case 0:
       case 4:
+        this.lastToken = startToken;
         return startToken;
 
       case 1:
-        this.prefixToken(startToken.key, depth + 1, jpath.concat('key'));
         this.prefixKey(startToken.key, depth, jpath);
 
         if(startToken.value !== null) {
-          this.prefixToken(startToken.value, depth + 1, jpath.concat('value'));
-
-          if(startToken.value.kind !== 3) {
-            this.prefixValue(startToken.value, depth);
-          }
+          this.prefixValue(startToken.value, depth, jpath);
         }
 
         return startToken;
@@ -84,17 +94,56 @@ class Refinery {
   /**
    * Adds the prefix to "key" tokens in a mapping. This is mostly whitespace.
    */
-  prefixKey(token: TokenRawValue | TokenCollection, depth: number, jpath: Array<string | number>): void {
+  prefixKey(token: TokenRawValue | TokenCollection, depth: number, jpath: JPath): void {
+    this.prefixToken(token, depth + 1, jpath.concat('key'));
+
     const rawToken: TokenRawValue = crawler.findFirstValueToken(token);
 
     if(!rawToken.prefix) {
       rawToken.prefix = [];
     }
 
-    // If there is no prefix AND this is not the very first key/value token.
-    if(!rawToken.prefix.length && jpath.join('.') !== 'mappings.0') {
-      const indent = `${this.indent.repeat(depth)}`;
-      rawToken.prefix.unshift(factory.createToken(indent));
+    const { prefix } = rawToken;
+    if(depth && prefix.length && REG_DASH.test(prefix[prefix.length - 1].value)) {
+      depth--;
+    }
+
+    // TODO: smarter indentation detection per token
+    // (using depth and leading whitespace in prefix)
+    const indent = this.indent.repeat(depth);
+    let missingIndent;
+    if(this.jsonDepth === -1) {
+      missingIndent = prefix.every(t =>
+        !REG_ALL_WHITESPACE.test(t.rawValue) && t.value.indexOf(indent) === -1
+      );
+    }
+    else {
+      const isFirstKey = rawToken.jpath[rawToken.jpath.length - 2] === 0;
+      const hasJsonStart = isFirstKey && prefix.some(t => REG_JSON_START.test(t.value));
+      const commaIndex = prefix.findIndex(t => REG_COMMA.test(t.value));
+
+      if(commaIndex === 0) {
+        prefix[0].value = prefix[0].rawValue = prefix[0].rawValue.replace(REG_LEADING_SPACE, '');
+      }
+
+      if(!isFirstKey && commaIndex === -1) {
+        prefix.push(factory.createToken(STR_COMMA));
+      }
+
+      missingIndent = commaIndex === -1 && !hasJsonStart;
+    }
+
+    // If there is no indent prefix AND this is not the very first key/value token.
+    if(missingIndent && jpath.join('.') !== STR_FIRST_TOKEN) {
+      prefix.forEach(t => {
+        if(REG_COMMENT.test(t.value)) {
+          t.value = t.rawValue = indent + t.rawValue.replace(REG_LEADING_SPACE, '');
+        }
+      });
+
+      if(!prefix.length || prefix[prefix.length - 1].rawValue.indexOf(STR_DASH) === -1) {
+        prefix.push(factory.createToken(`\n${indent}`));
+      }
     }
   }
 
@@ -102,56 +151,98 @@ class Refinery {
    * Adds the prefix to "value" tokens in a mapping. Most of the time
    * this will include a colon and white space.
    */
-  prefixValue(token: AnyToken, depth: number): void {
-    switch(token.kind) {
-      case 0:
-      case 4:
-        if (!token.prefix || !token.prefix.length) {
-          token.prefix = [ factory.createToken(`${STR_COLON} `) ];
-        }
+  prefixValue(token: AnyToken, depth: number, jpath: JPath): void {
+    this.prefixToken(token, depth + 1, jpath.concat('value'));
 
-        return;
+    const rawToken: TokenRawValue = crawler.findFirstValueToken(token);
 
-      case 2: {
-        const rawToken: TokenRawValue = crawler.findFirstValueToken(token);
+    if(!rawToken) {
+      // This only happens when an empty JSON object is used in yaml:
+      // foo: {}
+      return;
+    }
 
-        if(!rawToken.prefix) {
-          rawToken.prefix = [];
-        }
+    if(!rawToken.prefix) {
+      rawToken.prefix = [];
+    }
 
-        // only add the colon if it does not yet exist
-        if(rawToken.prefix.every(t => t.value.indexOf(STR_COLON) === -1)) {
-          rawToken.prefix.unshift(factory.createToken(STR_COLON));
-        }
+    this.addColonPrefix(rawToken.prefix);
+  }
 
-        return;
+  /**
+   * Given a prefix collection, adds the color if it does not exist
+   */
+  addColonPrefix(prefix: Array<TokenRawValue>) {
+    if(prefix.every(t => t.value.indexOf(STR_COLON) === -1)) {
+      let colon = STR_COLON;
+
+      if(!prefix.length || !/^\n/.test(prefix[prefix.length - 1].rawValue)) {
+        colon += ' ';
       }
 
-      case 3:
-        throw new Error('Must use addDashPrefix method');
-
-      default:
-        throw new Error(`Cannot add value prefix to token of kind: ${token.kind}`);
+      prefix.unshift(factory.createToken(colon));
     }
   }
 
   /**
    * Given a TokenMapping (kind: 2), refines each token in the mappings array.
    */
-  prefixMapping(startToken: TokenMapping, depth: number, jpath: Array<string | number>) {
-    startToken.mappings.forEach((token, i) => {
+  prefixMapping(startToken: TokenMapping, depth: number, jpath: JPath) {
+    let isJSON: boolean = false;
+    let firstToken: TokenRawValue;
+    let mapNeedsClosing: boolean = false;
+
+    startToken.mappings.forEach((token: TokenKeyValue, i: number) => {
       if (token === null) {
         return;
       }
 
+      const rawToken: TokenRawValue = crawler.findFirstValueToken(token);
+
+      if(!rawToken.prefix) {
+        rawToken.prefix = [];
+      }
+
+      if(!firstToken) {
+        firstToken = rawToken;
+
+        if(this.jsonDepth === -1) {
+          if(firstToken.prefix.find(t => REG_JSON_START.test(t.value))) {
+            // The object is a JSON-style object -> { foo: bar }
+            isJSON = true;
+            this.jsonDepth = depth;
+          }
+        }
+        else if(firstToken.prefix.every(t => !REG_JSON_START.test(t.value))) {
+          firstToken.prefix.unshift(factory.createToken(`${STR_OPEN_CURLY}\n${this.indent.repeat(depth)}`));
+          mapNeedsClosing = true;
+        }
+      }
+
       this.prefixToken(token, (token.kind === 2 ? depth + 1 : depth), jpath.concat('mappings', i));
     });
+
+    if(this.jsonDepth !== -1 && mapNeedsClosing) {
+      if(!startToken.suffix) {
+        startToken.suffix = [];
+      }
+      startToken.suffix.unshift(factory.createToken(`\n${this.indent.repeat(depth - 1)}${STR_CLOSE_CURLY}`));
+    }
+
+    // reset it back
+    if(isJSON) {
+      this.jsonDepth = -1;
+    }
   }
 
   /**
    * Given a TokenCollection (kind: 3), refines each token in the items array.
    */
-  prefixCollection(startToken: TokenCollection, depth: number, jpath: Array<string | number>) {
+  prefixCollection(startToken: TokenCollection, depth: number, jpath: JPath) {
+    let isJSON: boolean = false;
+    let firstToken: TokenRawValue;
+    let itemsNeedsClosing: boolean = false;
+
     startToken.items.forEach((token, i) => {
       if(token === null) {
         return;
@@ -165,18 +256,38 @@ class Refinery {
         rawToken.prefix = [];
       }
 
-      let dashIndex: number = rawToken.prefix.findIndex(t => t.value.indexOf('- ') !== -1);
+      if(!firstToken) {
+        firstToken = rawToken;
+
+        if(this.jsonDepth === -1) {
+          if(firstToken.prefix.find(t => REG_JSON_START.test(t.value))) {
+            // The object is a JSON-style array -> foo: [ bar ]
+            isJSON = true;
+            this.jsonDepth = depth;
+          }
+        }
+        else if(firstToken.prefix.every(t => !REG_JSON_START.test(t.value))) {
+          firstToken.prefix.unshift(factory.createToken(`${STR_OPEN_SQUARE}\n${this.indent.repeat(depth)}`));
+          itemsNeedsClosing = true;
+        }
+      }
+
+      let delimeterIndex: number = rawToken.prefix.findIndex(t =>
+        this.jsonDepth === -1
+          ? t.value.indexOf(STR_DASH) !== -1
+          : REG_JSON_START.test(t.value) || REG_COMMA.test(t.value)
+      );
 
       // if it's already there, no need to go further .
-      if(dashIndex !== -1) {
+      if(delimeterIndex !== -1) {
         return;
       }
 
       // First remove any whitespace tokens at the top of the prefix
-      let pre = rawToken.prefix[++dashIndex];
+      let pre = rawToken.prefix[ ++delimeterIndex] ;
       while(pre && REG_ALL_WHITESPACE.test(pre.rawValue)) {
-        rawToken.prefix.splice(dashIndex, 1);
-        pre = rawToken.prefix[dashIndex];
+        rawToken.prefix.splice(delimeterIndex, 1);
+        pre = rawToken.prefix[ delimeterIndex ];
       }
 
       /**
@@ -196,7 +307,21 @@ class Refinery {
       const firstIndex = lastTwo[1];
 
       while(lastTwo[0] === 'items' && lastTwo[1] >= firstIndex) {
-        const prefix = `\n${this.indent.repeat(depth - nesting)}- `;
+        let separator: string = '';
+        let prefix: string;
+
+        if(this.jsonDepth === -1) {
+          separator = nesting === 0 ? `${STR_DASH} ` : STR_DASH;
+          prefix = `\n${this.indent.repeat(depth - nesting)}${separator}`;
+        }
+        else {
+          if(lastTwo[1] > 0) {
+            separator = nesting === 0 ? `${STR_COMMA} ` : STR_COMMA;
+          }
+
+          prefix = `${separator}\n${this.indent.repeat(depth - nesting)}`;
+        }
+
         rawToken.prefix.unshift(factory.createToken(prefix));
 
         if(++nesting > 0 && firstIndex > 0) {
@@ -209,9 +334,22 @@ class Refinery {
 
       // the fist item in a collection should have a colon prefix
       if(i === 0 && rawToken.jpath[itemsIdx + 3] === 0) {
-        rawToken.prefix.unshift(factory.createToken(STR_COLON));
+        this.addColonPrefix(rawToken.prefix);
       }
     });
+
+    if(this.jsonDepth !== -1 && itemsNeedsClosing) {
+      if(!startToken.suffix) {
+        startToken.suffix = [];
+      }
+
+      startToken.suffix.unshift(factory.createToken(`\n${this.indent.repeat(depth - 1)}${STR_CLOSE_SQUARE}`));
+    }
+
+    // reset it back
+    if(isJSON) {
+      this.jsonDepth = -1;
+    }
   }
 
   reIndexToken(token: AnyToken, startPos: number): number {
@@ -225,25 +363,13 @@ class Refinery {
         token.startPosition = token.prefix.reduce((pos, prefix) => {
           return pos + prefix.rawValue.length;
         }, startPos);
+        this.newYaml += stringifier.stringifyToken(token);
         token.endPosition = token.startPosition + (token.rawValue || token.value).length;
         break;
 
       case 1:
         token.startPosition = startPos;
-        startPos = this.reIndexToken(token.key, startPos);
-
-        // Detect if this token was inserted at the end of the tree.
-        // If so, the old tail should become a prefix.
-        if(token.key.startPosition >= this.yaml.length - this.tail.length) {
-          const rawToken: TokenRawValue = crawler.findFirstValueToken(token.key);
-          const idx = rawToken.prefix[0].rawValue.indexOf(STR_COLON) === -1 ? 0 : 1;
-
-          // Remove any leading newlines from the insertion point
-          rawToken.prefix[idx].value = rawToken.prefix[idx].rawValue = rawToken.prefix[idx].rawValue.replace(DEFAULT_TAIL, '');
-          rawToken.prefix.splice(idx, 0, factory.createToken(`${this.tail}`));
-          this.tail = DEFAULT_TAIL;
-        }
-
+        startPos = this.reIndexToken(token.key, token.startPosition);
         token.endPosition = this.reIndexToken(token.value, startPos);
         break;
 
@@ -252,6 +378,10 @@ class Refinery {
         token.endPosition = token.mappings.reduce((pos, t) => {
           return this.reIndexToken(t, pos);
         }, startPos);
+
+        if(token.suffix) {
+          this.newYaml += token.suffix.reduce((s, t) => stringifier.stringifyToken(t, s), '');
+        }
         break;
 
       case 3:
@@ -259,6 +389,10 @@ class Refinery {
         token.endPosition = token.items.reduce((pos, t) => {
           return this.reIndexToken(t, pos);
         }, startPos);
+
+        if(token.suffix) {
+          this.newYaml += token.suffix.reduce((s, t) => stringifier.stringifyToken(t, s), '');
+        }
         break;
 
       default:
@@ -269,4 +403,4 @@ class Refinery {
   }
 }
 
-module.exports = Refinery;
+export default Refinery;
