@@ -4,7 +4,7 @@ import cx from 'classnames';
 
 import style from './style.css';
 
-export default class Notifications extends Component {
+class Notifications extends Component {
   static propTypes = {
     className: PropTypes.string,
     onRemove: PropTypes.func,
@@ -42,4 +42,6 @@ export default class Notifications extends Component {
       </div>
     );
   }
-};
+}
+
+export default Notifications;
