@@ -4,11 +4,48 @@ import { expect } from 'chai';
 import TokenSet from '../token-set';
 
 describe('TokenSet', () => {
-  describe('toYAML method', () => {
-    [ 'basic', 'simple', 'long', 'complex' ].forEach(file => {
-      it(`maintains source identity with ${file}.yaml`, () => {
-        const yaml = fs.readFileSync(path.join(__dirname, 'data', `${file}.yaml`), 'utf-8');
-        const set = new TokenSet(yaml);
+  [ 'basic', 'simple', 'long', 'complex' ].forEach(file => {
+    let yaml;
+    let set;
+
+    describe('Conventional YAML', () => {
+      const filename = `${file}.yaml`;
+
+      before(() => {
+        yaml = fs.readFileSync(path.join(__dirname, 'data', filename), 'utf-8');
+      });
+
+      beforeEach(() => {
+        set = new TokenSet(yaml);
+      });
+
+      it(`stringification maintains source identity with ${filename}`, () => {
+        expect(set.toYAML()).to.equal(yaml);
+      });
+
+      it(`refining maintains source identity with ${filename}`, () => {
+        set.refineTree();
+        expect(set.toYAML()).to.equal(yaml);
+      });
+    });
+
+    describe('JSON-like YAML', () => {
+      const filename = `${file}-json.yaml`;
+
+      before(() => {
+        yaml = fs.readFileSync(path.join(__dirname, 'data', filename), 'utf-8');
+      });
+
+      beforeEach(() => {
+        set = new TokenSet(yaml);
+      });
+
+      it(`stringification maintains source identity with ${filename}`, () => {
+        expect(set.toYAML()).to.equal(yaml);
+      });
+
+      it(`refining maintains source identity with ${filename}`, () => {
+        set.refineTree();
         expect(set.toYAML()).to.equal(yaml);
       });
     });
